Add lab4a

+++ lab3v/hash.h
@@ -1,7 +1,7 @@
 #ifndef LAB3_HASH_H
 #define LAB3_HASH_H
 
-int hash(const char *s);
+unsigned long long hash(const char *s);
 
 //int hash2(const char *s);
